+++ tetris_r4.vcxproj.filters
@@ -24,6 +24,12 @@
     <ClCompile Include="key_driver.c">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="draw.c">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="console.c">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="blocks.h">
@@ -36,6 +42,15 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="key_driver.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="sound.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="color.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="console.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
